Refactor authorization to authentication and align codebase
