+++ src/modules/auth/application/exceptions/auth-error.ts
