+++ src/modules/auth/application/guard/owner-role.guard.ts
@@ -6,14 +6,14 @@ import {
 } from '@nestjs/common';
 
 import { IResponse } from '@/common/response_service/interface/response.interface';
-import { Role } from '@/modules/authorization/domain/role.enum';
+import { Role } from '@/modules/auth/domain/role.enum';
 import { TeamService } from '@/modules/team/application/service/team.service';
 import { User } from '@/modules/user/domain/user.domain';
 
 import { AUTH_RESPONSE } from '../exceptions/auth-error';
 
 @Injectable()
-export class AdminRoleGuard implements CanActivate {
+export class OwnerRoleGuard implements CanActivate {
   constructor(private readonly teamService: TeamService) {}
 
   async canActivate(context: ExecutionContext) {
@@ -22,10 +22,10 @@ export class AdminRoleGuard implements CanActivate {
 
     const { payload: team } = await this.teamService.findOne(teamId);
 
-    const userReq: IResponse<User> = req.user;
+    const user: IResponse<User> = req.user;
 
     const isMemberOfATeam = team.userMembers.some(
-      (member) => member.userId === userReq.payload.id,
+      (member) => member.userId === user.payload.id,
     );
 
     if (!isMemberOfATeam) {
@@ -33,13 +33,13 @@ export class AdminRoleGuard implements CanActivate {
     }
 
     const userMember = team.userMembers.find(
-      (user) => user.userId === userReq.payload.id,
+      (member) => member.userId === user.payload.id,
     );
 
-    if (userMember.role === Role.OWNER || userMember.role === Role.ADMIN) {
-      return true;
+    if (userMember.role !== Role.OWNER) {
+      throw new UnauthorizedException(AUTH_RESPONSE.USER_ROLE_NOT_AUTHORIZED);
     }
 
-    throw new UnauthorizedException(AUTH_RESPONSE.USER_ROLE_NOT_AUTHORIZED);
+    return true;
   }
 }
