+++ src/modules/auth/domain/role.enum.ts
